requirements.txt
faker>=24.0.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0
pyarrow>=15.0.0   # necessário se output_format: parquet
